Add support for additional utput pre-processors, beyond just colors.
