+++ src/Joomla/Application/Cli/Output/Stdout.php
@@ -9,8 +9,7 @@
 namespace Joomla\Application\Cli\Output;
 
 use Joomla\Application\Cli\CliOutput;
-use Joomla\Application\Cli\ColorStyle;
-use Joomla\Application\Cli\ColorProcessor;
+use Joomla\Application\Cli\Output\Processors\ProcessorInterface;
 
 /**
  * Class Stdout.
@@ -19,46 +18,6 @@ use Joomla\Application\Cli\ColorProcessor;
  */
 class Stdout extends CliOutput
 {
-	/**
-	 * Constructor
-	 *
-	 * @since   1.0
-	 */
-	public function __construct()
-	{
-		$this->processor = new ColorProcessor;
-
-		$this->addPredefinedStyles();
-	}
-
-	/**
-	 * Set a processor
-	 *
-	 * @param   ColorProcessor  $processor  The color processor.
-	 *
-	 * @return  Stdout  Instance of $this to allow chaining.
-	 *
-	 * @since   1.0
-	 */
-	public function setProcessor(ColorProcessor $processor)
-	{
-		$this->processor = $processor;
-
-		return $this;
-	}
-
-	/**
-	 * Get a processor
-	 *
-	 * @return  ColorProcessor
-	 *
-	 * @since   1.0
-	 */
-	public function getProcessor()
-	{
-		return $this->processor;
-	}
-
 	/**
 	 * Write a string to standard output
 	 *
@@ -72,38 +31,6 @@ class Stdout extends CliOutput
 	public function out($text = '', $nl = true)
 	{
 		fwrite(STDOUT, $this->processor->process($text) . ($nl ? "\n" : null));
-
-		return $this;
-	}
-
-	/**
-	 * Adds predefined color styles to the ColorProcessor object
-	 *
-	 * @return  Stdout  Instance of $this to allow chaining.
-	 *
-	 * @since   1.0
-	 */
-	private function addPredefinedStyles()
-	{
-		$this->processor->addStyle(
-			'info',
-			new ColorStyle('green', '', array('bold'))
-		);
-
-		$this->processor->addStyle(
-			'comment',
-			new ColorStyle('yellow', '', array('bold'))
-		);
-
-		$this->processor->addStyle(
-			'question',
-			new ColorStyle('black', 'cyan')
-		);
-
-		$this->processor->addStyle(
-			'error',
-			new ColorStyle('white', 'red')
-		);
 
 		return $this;
 	}
